Moved Recurrence into Services, updated Use Case Diagram
